Fix missing logFinished method in Core
