+++ src/Eleventy.js
@@ -1,6 +1,4 @@
 import { relative } from "node:path";
-import chalk from "kleur";
-import { filesize } from "filesize";
 import debugUtil from "debug";
 
 import { TemplatePath } from "@11ty/eleventy-utils";
@@ -15,8 +13,6 @@ import TemplatePassthroughManager from "./TemplatePassthroughManager.js";
 import EleventyBaseError from "./Errors/EleventyBaseError.js";
 
 // Utils
-import simplePlural from "./Util/Pluralize.js";
-import { getEleventyPackageJson } from "./Util/ImportJsonSync.js";
 import checkPassthroughCopyBehavior from "./Util/PassthroughCopyBehaviorCheck.js";
 import PathPrefixer from "./Util/PathPrefixer.js";
 import PathNormalizer from "./Util/PathNormalizer.js";
@@ -25,7 +21,6 @@ import eventBus from "./EventBus.js";
 import { withResolvers } from "./Util/PromiseUtil.js";
 import GlobRemap from "./Util/GlobRemap.js";
 
-const pkg = getEleventyPackageJson();
 const debug = debugUtil("Eleventy");
 
 export default class Eleventy extends Core {
@@ -627,64 +622,6 @@ Arguments:
 	 */
 	getHelp() {
 		return Eleventy.getHelp();
-	}
-
-	/**
-	 * Logs some statistics after a complete run of Eleventy.
-	 *
-	 * @returns {string} ret - The log message.
-	 */
-	logFinished() {
-		if (!this.writer) {
-			throw new Error(
-				"Did you call Eleventy.init to create the TemplateWriter instance? Hint: you probably didn’t.",
-			);
-		}
-
-		let ret = [];
-
-		let {
-			copyCount,
-			copySize,
-			skipCount,
-			writeCount,
-			// renderCount, // files that render (costly) but may not write to disk
-		} = this.writer.getMetadata();
-
-		let slashRet = [];
-
-		if (copyCount) {
-			debug("Total passthrough copy aggregate size: %o", filesize(copySize));
-			slashRet.push(`Copied ${chalk.bold(copyCount)}`);
-		}
-
-		slashRet.push(
-			`Wrote ${chalk.bold(writeCount)} ${simplePlural(writeCount, "file", "files")}${
-				skipCount ? ` (skipped ${skipCount})` : ""
-			}`,
-		);
-
-		// slashRet.push(
-		// 	`${renderCount} rendered`
-		// )
-
-		if (slashRet.length) {
-			ret.push(slashRet.join(" "));
-		}
-
-		let time = (this.getNewTimestamp() - this.start) / 1000;
-		ret.push(
-			`in ${chalk.bold(time.toFixed(2))} ${simplePlural(time.toFixed(2), "second", "seconds")}`,
-		);
-
-		// More than 1 second total, show estimate of per-template time
-		if (time >= 1 && writeCount > 1) {
-			ret.push(`(${((time * 1000) / writeCount).toFixed(1)}ms each, v${pkg.version})`);
-		} else {
-			ret.push(`(v${pkg.version})`);
-		}
-
-		return ret.join(" ");
 	}
 }
